move filter under the packages directory
