+++ packages/filter/BUILD.bazel
@@ -1,0 +1,15 @@
+package(default_visibility = ["//visibility:public"])
+
+load("//tools/ts_project:index.bzl", "ts_project")
+
+ts_project(
+    name = "filter",
+    package_name = "@spica-server/filter",
+    srcs = glob([
+        "index.ts",
+        "src/*.ts",
+    ]),
+    deps = [
+        "//packages/database"
+    ],
+)
