+++ packages/filter/src/interface.ts
@@ -1,0 +1,11 @@
+export interface Extractor {
+  operators: string[];
+  factory: (expression: Expression) => string[][];
+}
+
+export interface Expression {
+  [key: string]: any;
+}
+
+export type KeyValidator = (key: string) => boolean;
+export type ValueConstructor<NewType = any> = (val) => NewType;
